refactor(vsock-guest): share worker ownership guards
